Fix a few harmless warnings.
git-svn-id: https://svn.filezilla-project.org/svn/FileZilla3/trunk@10727 ce1b9ea0-41a8-11dc-9b12-cb80f3b22977

+++ src/interface/listctrlex.cpp
@@ -876,7 +876,7 @@ int wxListCtrlEx::GetColumnActualIndex(int col) const
 	if (!m_pVisibleColumnMapping) {
 		return -1;
 	}
-	if (col < 0 || col >= m_columnInfo.size()) {
+	if (col < 0 || static_cast<size_t>(col) >= m_columnInfo.size()) {
 		return -1;
 	}
 	return m_pVisibleColumnMapping[col];
